+++ CHANGES.rst
@@ -41,6 +41,7 @@ describe the future plans.
    New Features
    ------------
 
+   * Add search for ESAFs & Proposals using Whoosh package.
    * Add support for direct access to read IS database.
    * Add Server class that chooses between DM or IS interface.
    * Integer timestamp PVs for ESAF start & end and Proposal start, end, & submitted.